async: Change VALIDATE_ASYNC opcode to GET_ASYNC_ITER
New opcode does the same things as GET_ITER, plus it validates that the
iter is either a generator object with CO_ASYNC flag, or, that the
iterable object had its __iter__ method tagged with __async__ attribute.

This opcode is supposed to be called instead of GET_ITER before
YIELD_FROM for 'await' calls.